requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Celery==5.3.0
redis==5.0.0
sendgrid==6.9.1
google-api-python-client==2.78.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.6.1
python-dotenv==1.0.0
requests==2.28.2
Flask-Cors==3.1.1
Flask-WTF==1.0.1
Flask-Login==0.6.2
